api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
motor==3.7.0
numpy==2.2.4
pandas==2.2.3
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.0
